usart: new method for USART assignment to edge
USARTS are now internally enumerated from A to D and their
base pointers are stored in an array indexed from 0 to 3.
Assignment of edges to USARTS is through symbolic name
USART_[A-D]. usart_init() needs to be called in order
A to D to match ISR assignments of rx/tx buffers.
Also some cleanups and removal of dead/commented code.