introduce "applet" mini-framework
Minimal framework for "applets" running from the main loop. A
new "struct applet" keeps pointers to an "init" and a "worker" function.
A "struct applet" instance can be registered with the framework by the
macro "applet_add", which puts an instance pointer into a dedicated ELF
section. Two functions applet_init_all() and applet_run_all() are then
called from main() to run the init and worker functions. This removes
compile and link time dependecies and avoids using #ifdef in order to run
a certain applet on the device.